Getting ready to move flexible schema to its own module.
