--- sqrl-base/src/main/java/com/datasqrl/model/schema/SchemaDefinition.java
@@ -1,5 +1,0 @@
-package com.datasqrl.model.schema;
-
-public interface SchemaDefinition {
-
-}
