+++ sqrl-execute/sqrl-execute-flink/sqrl-execute-flink-core/src/main/java/com/datasqrl/FlinkExecutablePlan.java
@@ -4,12 +4,10 @@ import com.datasqrl.config.BaseConnectorFactory;
 import com.datasqrl.io.formats.FormatFactory;
 import com.datasqrl.io.tables.TableSchemaFactory;
 import com.datasqrl.model.StreamType;
-import com.datasqrl.model.schema.SchemaDefinition;
 import com.datasqrl.serializer.SerializableSchema;
 import com.datasqrl.io.tables.TableConfig;
 import com.datasqrl.canonicalizer.NamePath;
 import com.datasqrl.model.LogicalStreamMetaData;
-import com.datasqrl.model.schema.TableDefinition;
 import com.fasterxml.jackson.annotation.JsonTypeInfo;
 import java.util.ArrayList;
 import java.util.List;
