--- sqrl-execute/sqrl-schema-avro/pom.xml
@@ -1,32 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-    <parent>
-        <groupId>com.datasqrl</groupId>
-        <artifactId>sqrl-execute</artifactId>
-        <version>0.1-SNAPSHOT</version>
-    </parent>
-
-    <artifactId>sqrl-schema-avro</artifactId>
-
-    <properties>
-        <maven.compiler.source>14</maven.compiler.source>
-        <maven.compiler.target>14</maven.compiler.target>
-        <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-    </properties>
-
-    <dependencies>
-        <dependency>
-            <groupId>com.datasqrl</groupId>
-            <artifactId>sqrl-base</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>com.datasqrl</groupId>
-            <artifactId>sqrl-io-core</artifactId>
-        </dependency>
-
-    </dependencies>
-
-</project>
