+++ sqrl-io/sqrl-io-core/src/main/java/com/datasqrl/engine/stream/RowMapper.java
@@ -1,7 +1,7 @@
 /*
  * Copyright (c) 2021, DataSQRL. All rights reserved. Use is subject to license terms.
  */
-package com.datasqrl.schema.converters;
+package com.datasqrl.engine.stream;
 
 import com.datasqrl.io.SourceRecord;
