+++ sqrl-io/sqrl-io-core/src/main/java/com/datasqrl/io/tables/SchemaValidator.java
@@ -1,7 +1,7 @@
 /*
  * Copyright (c) 2021, DataSQRL. All rights reserved. Use is subject to license terms.
  */
-package com.datasqrl.schema.input;
+package com.datasqrl.io.tables;
 
 //import com.datasqrl.engine.stream.FunctionWithError;
 import com.datasqrl.error.ErrorCollector;
