+++ sqrl-io/sqrl-io-core/src/main/java/com/datasqrl/io/tables/TableSchema.java
@@ -6,9 +6,8 @@ package com.datasqrl.io.tables;
 import com.datasqrl.canonicalizer.Name;
 import com.datasqrl.io.DataSystemConnectorSettings;
 import com.datasqrl.schema.converters.RowConstructor;
-import com.datasqrl.schema.converters.RowMapper;
+import com.datasqrl.engine.stream.RowMapper;
 import com.datasqrl.schema.input.SchemaAdjustmentSettings;
-import com.datasqrl.schema.input.SchemaValidator;
 
 public interface TableSchema {
   RowMapper getRowMapper(RowConstructor rowConstructor,
