+++ sqrl-io/sqrl-io-core/src/main/java/com/datasqrl/schema/converters/FlexibleSchemaRowMapper.java
@@ -3,8 +3,8 @@
  */
 package com.datasqrl.schema.converters;
 
+import com.datasqrl.engine.stream.RowMapper;
 import com.datasqrl.io.SourceRecord;
-import com.datasqrl.io.tables.TableSchema;
 import com.datasqrl.canonicalizer.Name;
 import com.datasqrl.schema.constraint.ConstraintHelper;
 import com.datasqrl.schema.input.FlexibleFieldSchema;
