+++ sqrl-io/sqrl-io-core/src/main/java/com/datasqrl/schema/input/external/AbstractElementDefinition.java
@@ -1,7 +1,7 @@
 /*
  * Copyright (c) 2021, DataSQRL. All rights reserved. Use is subject to license terms.
  */
-package com.datasqrl.model.schema;
+package com.datasqrl.schema.input.external;
 
 import java.io.Serializable;
