+++ sqrl-io/sqrl-io-core/src/main/java/com/datasqrl/schema/input/external/DatasetDefinition.java
@@ -3,7 +3,6 @@
  */
 package com.datasqrl.schema.input.external;
 
-import com.datasqrl.model.schema.TableDefinition;
 import java.io.Serializable;
 import java.util.List;
