+++ sqrl-io/sqrl-io-core/src/main/java/com/datasqrl/schema/input/external/SchemaExport.java
@@ -4,10 +4,6 @@
 package com.datasqrl.schema.input.external;
 
 import com.datasqrl.canonicalizer.SpecialName;
-import com.datasqrl.model.schema.AbstractElementDefinition;
-import com.datasqrl.model.schema.FieldDefinition;
-import com.datasqrl.model.schema.FieldTypeDefinitionImpl;
-import com.datasqrl.model.schema.TableDefinition;
 import com.datasqrl.schema.constraint.Constraint;
 import com.datasqrl.schema.input.FlexibleFieldSchema;
 import com.datasqrl.schema.input.FlexibleTableSchema;
