+++ sqrl-io/sqrl-io-core/src/main/java/com/datasqrl/schema/input/external/TableDefinition.java
@@ -1,12 +1,11 @@
 /*
  * Copyright (c) 2021, DataSQRL. All rights reserved. Use is subject to license terms.
  */
-package com.datasqrl.model.schema;
+package com.datasqrl.schema.input.external;
 
 import java.util.List;
 
-public class TableDefinition extends AbstractElementDefinition implements
-    SchemaDefinition {
+public class TableDefinition extends AbstractElementDefinition {
 
   public TableDefinition() {
