+++ sqrl-io/sqrl-io-core/src/main/java/com/datasqrl/schema/type/basic/IntervalType.java
