+++ sqrl-io/sqrl-io-schema-flexible/src/test/java/com/datasqrl/schema/type/BasicTypeTest.java
