+++ sqrl-planner/sqrl-planner-local/pom.xml
@@ -83,6 +83,16 @@
     </dependency>
     <dependency>
       <groupId>com.datasqrl</groupId>
+      <artifactId>sqrl-planner-schema-flexible</artifactId>
+      <scope>test</scope>
+    </dependency>
+    <dependency>
+      <groupId>com.datasqrl</groupId>
+      <artifactId>sqrl-io-schema-flexible</artifactId>
+      <scope>test</scope>
+    </dependency>
+    <dependency>
+      <groupId>com.datasqrl</groupId>
       <artifactId>sqrl-execute-flink-postgres</artifactId>
       <scope>test</scope>
     </dependency>
