+++ sqrl-planner/sqrl-planner-local/src/test/java/com/datasqrl/io/TestDataSetMonitoringIT.java
@@ -16,12 +16,7 @@ import com.datasqrl.discovery.TableWriter;
 import com.datasqrl.error.ErrorCollector;
 import com.datasqrl.io.tables.TableSource;
 import com.datasqrl.canonicalizer.Name;
-import com.datasqrl.metadata.MetadataStoreProvider;
-import com.datasqrl.schema.input.FlexibleTableSchema;
-import com.datasqrl.schema.input.FlexibleTableSchemaHolder;
 import com.datasqrl.schema.input.external.SchemaExport;
-import com.datasqrl.model.schema.TableDefinition;
-import com.datasqrl.util.FileTestUtil;
 import com.datasqrl.util.SnapshotTest;
 import com.datasqrl.util.TestDataset;
 import com.datasqrl.util.data.Nutshop;
@@ -29,9 +24,6 @@ import com.datasqrl.util.data.Retail;
 import com.datasqrl.util.data.Sensors;
 import com.datasqrl.util.junit.ArgumentProvider;
 
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.nio.file.StandardOpenOption;
 import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
