+++ sqrl-planner/sqrl-planner-schema-flexible/src/main/java/com/datasqrl/schema/converters/FlexibleTable2UTBConverter.java
