+++ sqrl-testing/sqrl-integration-tests/pom.xml
@@ -20,6 +20,16 @@
   <dependencies>
     <dependency>
       <groupId>com.datasqrl</groupId>
+      <artifactId>sqrl-planner-schema-flexible</artifactId>
+      <scope>test</scope>
+    </dependency>
+    <dependency>
+      <groupId>com.datasqrl</groupId>
+      <artifactId>sqrl-io-schema-flexible</artifactId>
+      <scope>test</scope>
+    </dependency>
+    <dependency>
+      <groupId>com.datasqrl</groupId>
       <artifactId>sqrl-discovery</artifactId>
       <scope>test</scope>
     </dependency>
